feat(304): refactor image metadata component
